requirements.txt
discord.py>=2.3,<3
python-dotenv>=1.0
orjson>=3.9
uvloop>=0.19; sys_platform != 'win32'
